refactor: extract web dashboard assets
