+++ internal/server/fleet_test.go
@@ -14,6 +14,7 @@ import (
 
 	"github.com/befeast/maestro/internal/config"
 	"github.com/befeast/maestro/internal/outcome"
+	"github.com/befeast/maestro/internal/server/web"
 	"github.com/befeast/maestro/internal/state"
 )
 
@@ -1366,7 +1367,7 @@ func TestFleetDashboard(t *testing.T) {
 	if w.Code != http.StatusOK {
 		t.Fatalf("status = %d, want %d", w.Code, http.StatusOK)
 	}
-	body := w.Body.String()
+	body := w.Body.String() + web.MustReadStatic("fleet.js") + web.MustReadStatic("fleet.css")
 	if ct := w.Header().Get("Content-Type"); ct != "text/html; charset=utf-8" {
 		t.Errorf("content-type = %q, want text/html", ct)
 	}
@@ -1675,7 +1676,7 @@ func fleetDashboardBodyWithProjects(t *testing.T, projects []FleetProject) strin
 	if w.Code != http.StatusOK {
 		t.Fatalf("status = %d, want %d", w.Code, http.StatusOK)
 	}
-	return w.Body.String()
+	return w.Body.String() + web.MustReadStatic("fleet.js") + web.MustReadStatic("fleet.css")
 }
 
 func fleetDashboardFixtureProjects(t *testing.T, count int) []FleetProject {
