+++ internal/server/server.go
@@ -17,6 +17,7 @@ import (
 
 	"github.com/befeast/maestro/internal/config"
 	"github.com/befeast/maestro/internal/outcome"
+	"github.com/befeast/maestro/internal/server/web"
 	"github.com/befeast/maestro/internal/state"
 	"github.com/befeast/maestro/internal/worker"
 )
@@ -50,6 +51,7 @@ func (s *Server) Start(ctx context.Context) error {
 	mux.HandleFunc("/api/v1/refresh", s.handleRefresh)
 	mux.HandleFunc("/api/v1/actions", s.handleAction)
 	mux.HandleFunc("/api/v1/", s.handleIssue)
+	mux.Handle("/static/", web.StaticHandler())
 	mux.HandleFunc("/", s.handleDashboard)
 
 	host := strings.TrimSpace(s.cfg.Server.Host)
@@ -925,6 +927,8 @@ func handleControlAction(w http.ResponseWriter, r *http.Request, readOnly bool, 
 	writeError(w, http.StatusNotImplemented, "approval-backed action endpoints are not implemented yet")
 }
 
+var dashboardHTML = web.MustReadTemplate("dashboard.html")
+
 func (s *Server) handleDashboard(w http.ResponseWriter, r *http.Request) {
 	if r.URL.Path != "/" {
 		http.NotFound(w, r)
@@ -940,763 +944,6 @@ func (s *Server) handleDashboard(w http.ResponseWriter, r *http.Request) {
 	page := strings.ReplaceAll(dashboardHTML, "__REPO_JSON__", string(repo))
 	fmt.Fprint(w, page)
 }
-
-const dashboardHTML = `<!DOCTYPE html>
-<html>
-<head>
-<title>maestro</title>
-<meta name="viewport" content="width=device-width, initial-scale=1">
-<style>
-  :root {
-    color-scheme: dark;
-    --bg: #0d1117;
-    --panel: #151b23;
-    --panel-2: #10161d;
-    --line: #29313d;
-    --text: #e6edf3;
-    --muted: #8b949e;
-    --accent: #58a6ff;
-    --ok: #3fb950;
-    --warn: #d29922;
-    --bad: #f85149;
-    --queued: #a371f7;
-  }
-  * { box-sizing: border-box; }
-  body {
-    margin: 0;
-    min-width: 320px;
-    background: var(--bg);
-    color: var(--text);
-    font: 14px/1.45 ui-sans-serif, system-ui, -apple-system, BlinkMacSystemFont, "Segoe UI", sans-serif;
-  }
-  header {
-    height: 56px;
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    gap: 20px;
-    padding: 0 18px;
-    border-bottom: 1px solid var(--line);
-    background: #0b1016;
-  }
-  h1 {
-    margin: 0;
-    font-size: 18px;
-    font-weight: 650;
-    letter-spacing: 0;
-  }
-  .repo { color: var(--muted); font-size: 13px; }
-  .topline, .stats, .workspace, .split, .log-head, .status-row {
-    display: flex;
-    align-items: center;
-    gap: 10px;
-  }
-  .topline { min-width: 0; }
-  .stats { flex-wrap: wrap; justify-content: flex-end; }
-  .stat {
-    min-width: 72px;
-    padding: 2px 4px;
-    text-align: right;
-  }
-  .stat strong { display: block; font-size: 15px; }
-  .stat span { display: block; color: var(--muted); font-size: 11px; }
-  main {
-    height: calc(100vh - 56px);
-    display: grid;
-    grid-template-columns: minmax(520px, 1fr) minmax(420px, 0.72fr);
-  }
-  section {
-    min-width: 0;
-    min-height: 0;
-    border-right: 1px solid var(--line);
-  }
-  section:last-child { border-right: 0; }
-  .toolbar {
-    height: 48px;
-    padding: 0 14px;
-    border-bottom: 1px solid var(--line);
-    background: var(--panel);
-    justify-content: space-between;
-  }
-  .toolbar h2 {
-    margin: 0;
-    font-size: 13px;
-    font-weight: 650;
-    color: var(--muted);
-    text-transform: uppercase;
-  }
-  .filter input {
-    width: min(260px, 36vw);
-    height: 30px;
-    border: 1px solid var(--line);
-    border-radius: 6px;
-    background: #0b1016;
-    color: var(--text);
-    padding: 0 10px;
-    outline: none;
-  }
-  .filter input:focus { border-color: var(--accent); }
-  .table-wrap {
-    height: calc(100% - 48px);
-    overflow: auto;
-  }
-  table {
-    width: 100%;
-    border-collapse: collapse;
-    table-layout: fixed;
-  }
-  th, td {
-    padding: 9px 10px;
-    border-bottom: 1px solid var(--line);
-    vertical-align: middle;
-    white-space: nowrap;
-    overflow: hidden;
-    text-overflow: ellipsis;
-  }
-  th {
-    position: sticky;
-    top: 0;
-    z-index: 1;
-    background: var(--panel);
-    color: var(--muted);
-    font-size: 12px;
-    font-weight: 600;
-    text-align: left;
-  }
-  tbody tr { cursor: pointer; }
-  tbody tr:hover, tbody tr.selected { background: #18212c; }
-  tbody tr.row-attention { background: rgba(248,81,73,.07); }
-  tbody tr.row-attention:hover, tbody tr.row-attention.selected { background: rgba(248,81,73,.13); }
-  a {
-    color: var(--accent);
-    text-decoration: none;
-  }
-  a:hover { text-decoration: underline; }
-  .slot { width: 92px; font-family: ui-monospace, SFMono-Regular, Menlo, Consolas, monospace; }
-  .issue { width: auto; }
-  .status { width: 128px; }
-  .backend { width: 102px; }
-  .pr { width: 64px; }
-  .runtime { width: 88px; }
-  .tokens { width: 78px; text-align: right; }
-  .pill {
-    display: inline-flex;
-    align-items: center;
-    max-width: 100%;
-    overflow: hidden;
-    height: 22px;
-    padding: 0 8px;
-    border-radius: 999px;
-    border: 1px solid var(--line);
-    font-size: 12px;
-    font-weight: 650;
-    text-overflow: ellipsis;
-    white-space: nowrap;
-  }
-  .issue-main {
-    display: flex;
-    align-items: baseline;
-    gap: 5px;
-    min-width: 0;
-    overflow: hidden;
-  }
-  .issue-main a { flex: 0 0 auto; }
-  .issue-title {
-    min-width: 0;
-    overflow: hidden;
-    text-overflow: ellipsis;
-    white-space: nowrap;
-  }
-  .s-running, .s-review_retry_running { color: var(--ok); border-color: rgba(63,185,80,.42); }
-  .s-pr_open { color: var(--warn); border-color: rgba(210,153,34,.48); }
-  .s-queued { color: var(--queued); border-color: rgba(163,113,247,.5); }
-  .s-review_retry_backoff, .s-review_retry_pending { color: var(--queued); border-color: rgba(163,113,247,.5); }
-  .s-review_retry_recheck { color: var(--warn); border-color: rgba(210,153,34,.48); }
-  .s-dead, .s-failed, .s-conflict_failed, .s-retry_exhausted { color: var(--bad); border-color: rgba(248,81,73,.5); }
-  .s-done { color: var(--muted); }
-  .pill-attention { color: var(--bad); border-color: rgba(248,81,73,.58); }
-  .log-panel {
-    min-width: 0;
-    display: grid;
-    grid-template-rows: 48px auto auto auto minmax(0, 1fr);
-    overflow: hidden;
-    background: #080c11;
-  }
-  .log-head {
-    padding: 0 14px;
-    border-bottom: 1px solid var(--line);
-    background: var(--panel);
-    justify-content: space-between;
-    min-width: 0;
-  }
-  .log-title { min-width: 0; font-weight: 650; }
-  .log-title span {
-    color: var(--muted);
-    font-weight: 500;
-    margin-left: 8px;
-  }
-  .log-meta {
-    color: var(--muted);
-    font-size: 12px;
-    white-space: nowrap;
-  }
-  .status-note {
-    display: none;
-    padding: 10px 14px;
-    border-bottom: 1px solid var(--line);
-    background: #0b1016;
-    color: var(--muted);
-    font-size: 12px;
-    max-height: 220px;
-    overflow: auto;
-  }
-  .status-note.visible { display: block; }
-  .status-note strong {
-    color: var(--text);
-    margin-right: 6px;
-  }
-  .status-note .links {
-    display: inline-flex;
-    gap: 10px;
-    margin-left: 10px;
-  }
-  .why-note {
-    color: var(--muted);
-    line-height: 1.45;
-    overflow-wrap: anywhere;
-  }
-  .supervisor-panel, .outcome-panel {
-	border-bottom: 1px solid var(--line);
-	background: #0b1016;
-	padding: 12px 14px;
-	font-size: 12px;
-	color: var(--muted);
-  }
-  .outcome-panel { background: #0f151c; }
-  .outcome-grid {
-    display: grid;
-    grid-template-columns: repeat(2, minmax(0, 1fr));
-    gap: 6px 12px;
-    margin-top: 6px;
-  }
-  .outcome-line { min-width: 0; overflow: hidden; text-overflow: ellipsis; white-space: nowrap; }
-  .outcome-line strong { color: var(--text); font-weight: 650; }
-  .supervisor-head, .supervisor-main, .supervisor-meta, .supervisor-links, .supervisor-actions {
-    display: flex;
-    align-items: center;
-    gap: 8px;
-    min-width: 0;
-  }
-  .supervisor-head { justify-content: space-between; margin-bottom: 6px; }
-  .supervisor-title { color: var(--text); font-weight: 650; }
-  .supervisor-time { white-space: nowrap; }
-  .supervisor-action {
-    color: var(--text);
-    font-weight: 650;
-    overflow: hidden;
-    text-overflow: ellipsis;
-    white-space: nowrap;
-  }
-  .supervisor-summary { margin-top: 4px; color: var(--text); }
-  .supervisor-meta, .supervisor-links, .supervisor-actions { flex-wrap: wrap; margin-top: 6px; }
-  .supervisor-reasons {
-    margin: 6px 0 0;
-    padding-left: 16px;
-  }
-  .supervisor-reasons li { margin: 2px 0; }
-  .supervisor-approval, .action-btn {
-    height: 24px;
-    max-width: 150px;
-    border: 1px solid rgba(210,153,34,.45);
-    border-radius: 6px;
-    background: rgba(210,153,34,.08);
-    color: var(--warn);
-    cursor: not-allowed;
-    overflow: hidden;
-    padding: 0 8px;
-    text-overflow: ellipsis;
-    white-space: nowrap;
-  }
-  .worker-actions {
-    display: grid;
-    gap: 6px;
-    margin-top: 8px;
-  }
-  .worker-actions span { color: var(--text); font-weight: 650; }
-  .action-list { display: inline-flex; gap: 6px; flex-wrap: wrap; }
-  .action-item { min-width: 134px; max-width: 210px; }
-  .action-detail {
-    display: grid;
-    gap: 2px;
-    margin-top: 4px;
-    color: var(--muted);
-    font-size: 11px;
-    line-height: 1.3;
-  }
-  .action-detail strong { color: var(--text); font-weight: 650; }
-  .action-note { margin-top: 6px; color: var(--muted); }
-  .supervisor-empty { color: var(--muted); }
-  pre {
-    margin: 0;
-    min-height: 0;
-    overflow: auto;
-    padding: 14px;
-    color: #d1d7e0;
-    font: 12px/1.5 ui-monospace, SFMono-Regular, Menlo, Consolas, monospace;
-    white-space: pre-wrap;
-    word-break: break-word;
-  }
-  .empty {
-    padding: 24px;
-    color: var(--muted);
-  }
-  .muted { color: var(--muted); }
-  .error { color: var(--bad); }
-  @media (max-width: 980px) {
-    header { height: auto; align-items: flex-start; flex-direction: column; padding: 12px 14px; }
-    .stats { justify-content: flex-start; }
-    main { height: auto; min-height: calc(100vh - 112px); grid-template-columns: 1fr; }
-    section { border-right: 0; border-bottom: 1px solid var(--line); }
-    .table-wrap { height: 52vh; }
-    .log-panel { height: 48vh; }
-    .backend, .tokens { display: none; }
-  }
-</style>
-</head>
-<body>
-<header>
-  <div class="topline">
-    <h1>Maestro</h1>
-    <div class="repo" id="repo"></div>
-  </div>
-  <div class="stats" id="stats"></div>
-</header>
-<main>
-  <section>
-    <div class="toolbar workspace">
-      <h2>Workers</h2>
-      <label class="filter"><input id="filter" type="search" placeholder="Filter"></label>
-    </div>
-    <div class="table-wrap">
-      <table>
-        <thead>
-          <tr>
-            <th class="slot">Slot</th>
-            <th class="issue">Issue</th>
-            <th class="status">Status</th>
-            <th class="backend">Backend</th>
-            <th class="pr">PR</th>
-            <th class="runtime">Runtime</th>
-            <th class="tokens">Tokens</th>
-          </tr>
-        </thead>
-        <tbody id="workers"></tbody>
-      </table>
-    </div>
-  </section>
-  <section class="log-panel">
-    <div class="log-head">
-      <div class="log-title" id="log-title">Log <span></span></div>
-      <div class="log-meta" id="log-meta"></div>
-    </div>
-	<div class="supervisor-panel" id="supervisor-panel"></div>
-	<div class="outcome-panel" id="outcome-panel"></div>
-	<div class="status-note" id="status-note"></div>
-    <pre id="log"><span class="muted">Select a worker.</span></pre>
-  </section>
-</main>
-<script>
-window.MAESTRO_REPO = __REPO_JSON__;
-
-const state = {
-  workers: [],
-  supervisor: null,
-  outcome: null,
-  selected: "",
-  filter: "",
-  lastLog: null
-};
-
-const statusRank = {
-  review_retry_running: 0,
-  running: 0,
-  review_retry_recheck: 1,
-  pr_open: 1,
-  review_retry_pending: 2,
-  review_retry_backoff: 2,
-  queued: 2,
-  dead: 3,
-  failed: 4,
-  conflict_failed: 5,
-  retry_exhausted: 6,
-  done: 7
-};
-
-const repoEl = document.getElementById("repo");
-const statsEl = document.getElementById("stats");
-const workersEl = document.getElementById("workers");
-const filterEl = document.getElementById("filter");
-const logEl = document.getElementById("log");
-const logTitleEl = document.getElementById("log-title");
-const logMetaEl = document.getElementById("log-meta");
-const statusNoteEl = document.getElementById("status-note");
-const supervisorPanelEl = document.getElementById("supervisor-panel");
-const outcomePanelEl = document.getElementById("outcome-panel");
-
-repoEl.textContent = window.MAESTRO_REPO || "";
-
-filterEl.addEventListener("input", () => {
-  state.filter = filterEl.value.toLowerCase();
-  renderWorkers();
-});
-
-function escapeText(value) {
-  return String(value ?? "").replace(/[&<>"']/g, ch => ({
-    "&": "&amp;",
-    "<": "&lt;",
-    ">": "&gt;",
-    "\"": "&quot;",
-    "'": "&#39;"
-  }[ch]));
-}
-
-function compactNumber(value) {
-  const n = Number(value || 0);
-  if (!n) return "-";
-  if (n < 1000) return String(n);
-  if (n < 1000000) return (n / 1000).toFixed(n < 10000 ? 1 : 0).replace(/\.0$/, "") + "k";
-  return (n / 1000000).toFixed(1).replace(/\.0$/, "") + "M";
-}
-
-function linkHTML(url, label) {
-  if (!url) return escapeText(label);
-  return '<a href="' + escapeText(url) + '" target="_blank" rel="noreferrer">' + escapeText(label) + '</a>';
-}
-
-function issueSummaryHTML(worker) {
-  const issue = worker.issue_number ? "#" + worker.issue_number : "-";
-  return '<span class="issue-main">' + linkHTML(worker.issue_url, issue) +
-    '<span class="issue-title">' + escapeText(worker.issue_title || "") + '</span></span>';
-}
-
-function issueSummaryText(worker) {
-  const issue = worker.issue_number ? "#" + worker.issue_number : "-";
-  return (issue + " " + (worker.issue_title || "")).trim();
-}
-
-function actionLabel(action) {
-  return String(action || "-").replace(/_/g, " ");
-}
-
-function actionDisabledReason(actions) {
-  const action = (actions || []).find(item => item.disabled_reason);
-  return action ? action.disabled_reason : "Write actions require approval-backed configuration.";
-}
-
-function actionTargetText(action) {
-  const parts = [];
-  if (action.target) parts.push(action.target);
-  if (action.issue_number) parts.push("issue #" + action.issue_number);
-  if (action.pr_number) parts.push("PR #" + action.pr_number);
-  return parts.length ? parts.join(" · ") : "project";
-}
-
-function actionPolicyText(action) {
-  if (action.approval_policy) return actionLabel(action.approval_policy);
-  return action.requires_approval ? "manual approval required" : "none";
-}
-
-function actionDetailHTML(action) {
-  const description = action.description ? '<div><strong>Would</strong> ' + escapeText(action.description.replace(/^Would\s+/i, "")) + '</div>' : "";
-  return '<div class="action-detail">' + description +
-    '<div><strong>Scope</strong> ' + escapeText(actionLabel(action.scope || "unknown")) + '</div>' +
-    '<div><strong>Target</strong> ' + escapeText(actionTargetText(action)) + '</div>' +
-    '<div><strong>Approval</strong> ' + escapeText(actionPolicyText(action)) + '</div>' +
-    '<div><strong>Disabled</strong> ' + escapeText(action.disabled_reason || "Write action unavailable") + '</div>' +
-    '</div>';
-}
-
-function renderActionButtons(actions, showDetails) {
-  const items = actions || [];
-  if (!items.length) return "";
-  return '<div class="action-list">' + items.map(action =>
-    '<div class="action-item"><button type="button" class="action-btn" disabled aria-disabled="true" title="' +
-    escapeText(action.disabled_reason || "Write action unavailable") + '">' +
-    escapeText(action.label || actionLabel(action.id)) + '</button>' +
-    (showDetails ? actionDetailHTML(action) : "") + '</div>'
-  ).join("") + '</div>';
-}
-
-function renderWorkerActions(actions) {
-  if (!actions || !actions.length) return "";
-  return '<div class="worker-actions"><span>Actions</span>' + renderActionButtons(actions, false) + '</div>' +
-    '<div class="action-note">' + escapeText(actionDisabledReason(actions)) + '</div>';
-}
-
-function formatTimestamp(value) {
-  if (!value) return "-";
-  const date = new Date(value);
-  if (Number.isNaN(date.getTime())) return String(value);
-  const seconds = Math.max(0, Math.round((Date.now() - date.getTime()) / 1000));
-  let relative = seconds + "s ago";
-  if (seconds >= 86400) relative = Math.floor(seconds / 86400) + "d ago";
-  else if (seconds >= 3600) relative = Math.floor(seconds / 3600) + "h ago";
-  else if (seconds >= 60) relative = Math.floor(seconds / 60) + "m ago";
-  return date.toLocaleString() + " (" + relative + ")";
-}
-
-function targetLinksHTML(links) {
-  return (links || []).map(link => linkHTML(link.url, link.label)).join("");
-}
-
-function queueText(queue) {
-  if (!queue || !queue.enabled) return "";
-  const label = queue.label || "Queue";
-  if (queue.position && queue.total) return label + ": " + queue.position + " of " + queue.total;
-  if (queue.total) return label + ": " + queue.total + " item" + (queue.total === 1 ? "" : "s");
-  return label + ": empty";
-}
-
-function displayStatus(worker) {
-  return worker.display_status || worker.status || "-";
-}
-
-function statusLabel(worker) {
-  if (worker.status === "running" && worker.alive === false) return "running stale";
-  return displayStatus(worker);
-}
-
-function pillClass(worker) {
-  const base = "pill s-" + escapeText(displayStatus(worker) || "unknown");
-  if (worker.needs_attention || (worker.status === "running" && worker.alive === false)) {
-    return base + " pill-attention";
-  }
-  return base;
-}
-
-function workerMatches(worker) {
-  if (!state.filter) return true;
-  const text = [worker.slot, worker.issue_number, worker.issue_title, worker.status, displayStatus(worker), worker.backend, worker.pr_number]
-    .join(" ")
-    .toLowerCase();
-  return text.includes(state.filter);
-}
-
-function sortWorkers(workers) {
-  return [...workers].sort((a, b) => {
-    const ar = statusRank[displayStatus(a)] ?? 99;
-    const br = statusRank[displayStatus(b)] ?? 99;
-    if (ar !== br) return ar - br;
-    return String(b.started_at || "").localeCompare(String(a.started_at || ""));
-  });
-}
-
-function renderStats(summary, total, maxParallel, readOnly) {
-  const running = (summary.running || 0) + (summary.review_retry_running || 0);
-  const prOpen = (summary.pr_open || 0) + (summary.review_retry_recheck || 0);
-  const failed = (summary.dead || 0) + (summary.failed || 0) + (summary.retry_exhausted || 0) + (summary.conflict_failed || 0);
-  const items = [
-    ["Running", running + " / " + maxParallel],
-    ["PR open", prOpen],
-    ["Failed", failed],
-    ["Sessions", total],
-    ["Mode", readOnly ? "Read-only" : "Control"]
-  ];
-  statsEl.innerHTML = items.map(([label, value]) =>
-    '<div class="stat"><strong>' + escapeText(value) + '</strong><span>' + escapeText(label) + '</span></div>'
-  ).join("");
-}
-
-function renderSupervisor(info) {
-  if (!info || !info.has_run || !info.latest) {
-    const empty = info && info.empty_state ? info.empty_state : "No Supervisor has run yet.";
-    supervisorPanelEl.innerHTML = '<div class="supervisor-head">' +
-      '<span class="supervisor-title">Supervisor</span>' +
-      '<span class="supervisor-time">empty</span>' +
-      '</div>' +
-      '<div class="supervisor-empty">' + escapeText(empty) + '</div>';
-    return;
-  }
-
-  const latest = info.latest;
-  const links = targetLinksHTML(latest.target_links);
-  const queue = queueText(latest.queue);
-  const meta = [
-    latest.risk ? "Risk " + latest.risk : "",
-    latest.confidence ? "Confidence " + Number(latest.confidence).toFixed(2) : "",
-    queue
-  ].filter(Boolean);
-  const reasons = (latest.stuck_reasons && latest.stuck_reasons.length ? latest.stuck_reasons : latest.reasons || []).slice(0, 3);
-  const reasonHTML = reasons.length ? '<ul class="supervisor-reasons">' + reasons.map(reason =>
-    '<li>' + escapeText(reason) + '</li>'
-  ).join("") + '</ul>' : "";
-  const lastSafe = info.last_safe_action ? '<div class="supervisor-meta">' +
-    '<span>Last safe action: ' + escapeText(actionLabel(info.last_safe_action.action)) + '</span>' +
-    '<span>' + escapeText(formatTimestamp(info.last_safe_action.created_at)) + '</span>' +
-    '</div>' : "";
-  const approvals = (info.approval_actions || []).length ? '<div class="supervisor-actions">' +
-    '<span>Requires approval:</span>' +
-    (info.approval_actions || []).map(action =>
-      '<button class="supervisor-approval" disabled title="' + escapeText(action.disabled_reason || "Controls not available yet") + '">' +
-      escapeText(actionLabel(action.action)) +
-      '</button>'
-    ).join("") +
-    '</div>' : "";
-
-  supervisorPanelEl.innerHTML = '<div class="supervisor-head">' +
-    '<span class="supervisor-title">Supervisor</span>' +
-    '<span class="supervisor-time">' + escapeText(formatTimestamp(latest.created_at)) + '</span>' +
-    '</div>' +
-    '<div class="supervisor-main">' +
-    '<span class="supervisor-action">' + escapeText(actionLabel(latest.recommended_action)) + '</span>' +
-    (links ? '<span class="supervisor-links">' + links + '</span>' : "") +
-    '</div>' +
-    (latest.summary ? '<div class="supervisor-summary">' + escapeText(latest.summary) + '</div>' : "") +
-    (meta.length ? '<div class="supervisor-meta">' + meta.map(item => '<span>' + escapeText(item) + '</span>').join("") + '</div>' : "") +
-    reasonHTML + lastSafe + approvals;
-}
-
-function renderOutcome(outcome) {
-  const o = outcome || {};
-  const configured = o.configured === true;
-  const goal = configured ? (o.goal || "Configured outcome") : "No outcome brief configured";
-  const target = o.runtime_target || "-";
-  const health = o.health_state || (configured ? "unknown" : "not_configured");
-  const next = o.next_action || (configured ? "Verify runtime health." : "Add an outcome brief to config.");
-  const host = o.runtime_host ? " · " + o.runtime_host : "";
-  const checked = o.health_checked_at ? formatTimestamp(o.health_checked_at) : "-";
-  const summary = o.health_summary || "";
-  outcomePanelEl.innerHTML = '<div class="supervisor-head">' +
-    '<span class="supervisor-title">Outcome</span>' +
-    '<span class="supervisor-time">' + escapeText(health.replace(/_/g, " ")) + '</span>' +
-    '</div>' +
-    '<div class="outcome-grid">' +
-      '<div class="outcome-line" title="' + escapeText(goal) + '"><strong>Goal</strong> ' + escapeText(goal) + '</div>' +
-      '<div class="outcome-line" title="' + escapeText(target + host) + '"><strong>Runtime</strong> ' + escapeText(target + host) + '</div>' +
-      '<div class="outcome-line" title="' + escapeText(next) + '"><strong>Next</strong> ' + escapeText(next) + '</div>' +
-      '<div class="outcome-line"><strong>Health</strong> ' + escapeText(health.replace(/_/g, " ")) + '</div>' +
-      '<div class="outcome-line"><strong>Checked</strong> ' + escapeText(checked) + '</div>' +
-      (summary ? '<div class="outcome-line" title="' + escapeText(summary) + '"><strong>Signal</strong> ' + escapeText(summary) + '</div>' : "") +
-    '</div>';
-}
-
-function renderWorkers() {
-  const visible = sortWorkers(state.workers).filter(workerMatches);
-  if (visible.length === 0) {
-    workersEl.innerHTML = '<tr><td colspan="7" class="empty">No workers.</td></tr>';
-    return;
-  }
-  workersEl.innerHTML = visible.map(worker => {
-    const selected = worker.slot === state.selected ? " selected" : "";
-    const attention = worker.needs_attention ? " row-attention" : "";
-    const pr = worker.pr_number ? "#" + worker.pr_number : "-";
-    return '<tr class="' + selected + attention + '" data-slot="' + escapeText(worker.slot) + '">' +
-      '<td class="slot">' + escapeText(worker.slot) + '</td>' +
-      '<td class="issue" title="' + escapeText(issueSummaryText(worker)) + '">' + issueSummaryHTML(worker) + '</td>' +
-      '<td class="status"><span class="' + pillClass(worker) + '">' + escapeText(statusLabel(worker)) + '</span></td>' +
-      '<td class="backend">' + escapeText(worker.backend || "-") + '</td>' +
-      '<td class="pr">' + linkHTML(worker.pr_url, pr) + '</td>' +
-      '<td class="runtime">' + escapeText(worker.runtime || "-") + '</td>' +
-      '<td class="tokens">' + compactNumber(worker.tokens_used_total) + '</td>' +
-    '</tr>';
-  }).join("");
-  workersEl.querySelectorAll("tr[data-slot]").forEach(row => {
-    row.addEventListener("click", () => selectWorker(row.dataset.slot));
-  });
-}
-
-function selectWorker(slot) {
-  state.selected = slot;
-  state.lastLog = null;
-  renderWorkers();
-  renderSelectedDetails();
-  loadLog();
-}
-
-function emptyLogText(worker) {
-  if (!worker) return "No log output yet.";
-  if (worker.status === "running" && worker.backend === "claude") {
-    return "No log output yet. Claude print mode may stay quiet until it finishes.";
-  }
-  if (worker.status === "running") return "No log output yet. Worker is still running.";
-  return "No log output.";
-}
-
-function renderSelectedDetails() {
-  const worker = state.workers.find(item => item.slot === state.selected);
-  if (!worker) {
-    statusNoteEl.classList.remove("visible");
-    statusNoteEl.innerHTML = "";
-    return;
-  }
-
-  const links = [];
-  if (worker.issue_url) links.push(linkHTML(worker.issue_url, "Issue #" + worker.issue_number));
-  if (worker.pr_url) links.push(linkHTML(worker.pr_url, "PR #" + worker.pr_number));
-  const retry = worker.next_retry_at ? " Next retry: " + worker.next_retry_at + "." : "";
-  const next = worker.next_action ? " Next: " + worker.next_action : "";
-  statusNoteEl.innerHTML = '<div class="why-note"><strong>Why</strong>' +
-    escapeText((worker.status_reason || "Waiting for next reconciliation cycle.") + retry + next) +
-    (links.length ? '<span class="links">' + links.join("") + '</span>' : "") + '</div>' +
-    renderWorkerActions(worker.actions || []);
-  statusNoteEl.classList.add("visible");
-}
-
-async function loadState() {
-  try {
-    const response = await fetch("/api/v1/state", { cache: "no-store" });
-    if (!response.ok) throw new Error(await response.text());
-    const data = await response.json();
-	state.workers = data.all || [];
-	state.supervisor = data.supervisor || null;
-	state.outcome = data.outcome || null;
-	renderStats(data.summary || {}, state.workers.length, data.max_parallel || 0, data.read_only);
-	renderSupervisor(state.supervisor);
-	renderOutcome(state.outcome);
-    if (!state.selected && state.workers.length) state.selected = sortWorkers(state.workers)[0].slot;
-    if (state.selected && !state.workers.some(worker => worker.slot === state.selected)) {
-      state.selected = state.workers.length ? sortWorkers(state.workers)[0].slot : "";
-      state.lastLog = null;
-    }
-    renderWorkers();
-    renderSelectedDetails();
-  } catch (err) {
-    statsEl.innerHTML = '<span class="error">' + escapeText(err.message) + '</span>';
-  }
-}
-
-async function loadLog() {
-  if (!state.selected) {
-    logTitleEl.innerHTML = 'Log <span></span>';
-    logMetaEl.textContent = "";
-    logEl.textContent = "Select a worker.";
-    renderSelectedDetails();
-    return;
-  }
-  const worker = state.workers.find(item => item.slot === state.selected);
-  logTitleEl.innerHTML = 'Log <span>' + escapeText(state.selected) + (worker ? " #" + escapeText(worker.issue_number) : "") + '</span>';
-  try {
-    const response = await fetch("/api/v1/logs/" + encodeURIComponent(state.selected) + "?lines=260", { cache: "no-store" });
-    if (!response.ok) throw new Error(await response.text());
-    const data = await response.json();
-    const text = data.text || "";
-    logMetaEl.textContent = (data.truncated ? "tail " : "") + (data.updated_at || "");
-    if (text !== state.lastLog) {
-      state.lastLog = text;
-      logEl.textContent = text || emptyLogText(worker);
-      logEl.scrollTop = logEl.scrollHeight;
-    }
-  } catch (err) {
-    logEl.textContent = "Log error: " + err.message;
-  }
-}
-
-loadState().then(loadLog);
-setInterval(loadState, 3000);
-setInterval(loadLog, 2000);
-</script>
-</body>
-</html>`
 
 func escapeHTML(s string) string {
 	s = strings.ReplaceAll(s, "&", "&amp;")
