+++ internal/server/web/templates/fleet.html
@@ -1,0 +1,135 @@
+<!DOCTYPE html>
+<html>
+<head>
+<title>maestro fleet</title>
+<meta name="viewport" content="width=device-width, initial-scale=1">
+<link rel="stylesheet" href="/static/fleet.css">
+
+</head>
+<body data-page="fleet">
+<header>
+  <div>
+    <h1>Maestro Fleet</h1>
+    <div class="sub" id="subtitle">Loading projects...</div>
+    <div class="fleet-verdict verdict-healthy" id="fleet-verdict">Loading supervisor heartbeat...</div>
+  </div>
+  <div class="stats" id="stats"></div>
+</header>
+<main>
+  <section class="project-rail" id="project-rail" aria-live="polite">
+    <div class="section-head">
+      <div>
+        <h2>Project Rail</h2>
+        <div class="sub">All configured projects in one scan-friendly view: state, queue, PRs, outcome, freshness, and links.</div>
+      </div>
+      <div class="section-note" id="project-rail-summary">{{FLEET_PROJECT_RAIL_SUMMARY}}</div>
+    </div>
+    <div class="project-rail-controls">
+      <label for="project-filter"><span>Project Search</span><input id="project-filter" type="search" placeholder="Filter by project, repo, state, queue, or outcome"></label>
+      <div class="project-rail-help">Search narrows the rail only; the rail remains the primary fleet overview.</div>
+    </div>
+    <div class="project-rail-scroll">
+      <table class="project-rail-table" aria-label="Configured project fleet rail">
+        <thead>
+          <tr>
+            <th class="project-rail-project">Project</th>
+            <th class="project-rail-state-cell">State</th>
+            <th class="project-rail-queue-cell">Queue</th>
+            <th class="project-rail-pr-cell">PR</th>
+            <th class="project-rail-outcome-cell">Outcome</th>
+            <th class="project-rail-freshness-cell">Last activity</th>
+            <th class="project-rail-links-cell">Links/actions</th>
+          </tr>
+        </thead>
+        <tbody id="project-rail-body">{{FLEET_PROJECT_RAIL_ROWS}}</tbody>
+      </table>
+    </div>
+  </section>
+  <details class="project-overview project-diagnostics" id="project-diagnostics">
+    <summary>Project diagnostics and raw state drilldown</summary>
+    <div class="section-head">
+      <div>
+        <h2>Project Diagnostics</h2>
+        <div class="sub">Expanded per-project cards retain raw queue, supervisor, approval, and read-only control diagnostics.</div>
+      </div>
+      <div class="section-note" id="project-summary">Loading project diagnostics...</div>
+    </div>
+    <div class="grid" id="projects"></div>
+  </details>
+  <section class="approval-inbox" id="approval-inbox" aria-live="polite">
+    <div class="section-head">
+      <div>
+        <h2>Approval Inbox</h2>
+        <div class="sub">Read-only lifecycle view of supervisor approvals across projects.</div>
+      </div>
+      <div class="section-note" id="approval-summary">Loading approvals...</div>
+    </div>
+    <div class="approval-list" id="approval-list"></div>
+  </section>
+  <section class="attention-inbox" id="attention-inbox" aria-live="polite">
+    <div class="section-head">
+      <div>
+        <h2>Attention Inbox</h2>
+        <div class="sub">One ordered list of the worker/project items that need an operator now.</div>
+      </div>
+      <div class="section-note" id="attention-summary">Loading attention...</div>
+    </div>
+    <div class="attention-list" id="attention-list"></div>
+  </section>
+  <section class="fleet-workers">
+    <div class="section-head">
+      <div>
+        <h2>Fleet Workers</h2>
+        <div class="sub">Operator-focused worker queue. Switch scope when you need history.</div>
+      </div>
+      <div class="worker-scope-status">
+        <div class="section-note" id="worker-summary">Loading workers...</div>
+        <button type="button" id="worker-project-reset" class="scope-reset-button" hidden>Show all projects</button>
+      </div>
+    </div>
+    <div class="worker-controls" id="worker-controls">
+      <label class="search-control" for="worker-filter"><span>Search</span><input id="worker-filter" type="search" placeholder="Project, issue, status, backend, or PR"></label>
+      <label for="scope-filter"><span>Scope</span><select id="scope-filter"><option value="operator">Needs action/live</option><option value="attention">Attention only</option><option value="live">Live only</option><option value="recent">Done/history</option><option value="all">All workers</option></select></label>
+      <label for="status-filter"><span>Status</span><select id="status-filter"></select></label>
+      <label for="backend-filter"><span>Backend</span><select id="backend-filter"></select></label>
+      <label for="pr-filter"><span>PR</span><select id="pr-filter"><option value="all">Any PR</option><option value="with">With PR</option><option value="without">No PR</option></select></label>
+      <label for="worker-sort"><span>Sort</span><select id="worker-sort"><option value="status">Status</option><option value="project">Project</option><option value="issue">Issue</option><option value="runtime">Runtime</option><option value="pr">PR</option></select></label>
+      <button type="button" id="sort-direction" aria-label="Toggle sort direction">Asc</button>
+    </div>
+    <div class="table-scroll">
+      <table class="worker-table">
+        <thead>
+          <tr>
+            <th class="project-col">Project</th>
+            <th class="slot-col">Slot</th>
+            <th class="issue-col">Issue</th>
+            <th class="status-col">Status</th>
+            <th class="backend-col">Backend</th>
+            <th class="pr-col">PR</th>
+            <th class="runtime-col">Runtime</th>
+            <th class="tokens-col">Tokens</th>
+            <th class="action-col">Actions</th>
+          </tr>
+        </thead>
+        <tbody id="fleet-workers-body"></tbody>
+      </table>
+    </div>
+  </section>
+  <section class="worker-detail" id="worker-detail" aria-live="polite">
+    <div class="section-head">
+      <div>
+        <h2>Worker Detail</h2>
+        <div class="sub">Select a worker row to inspect session state and recent log output.</div>
+      </div>
+      <div class="section-note" id="worker-detail-summary">No worker selected</div>
+    </div>
+    <div class="detail-body" id="worker-detail-body">
+      <div class="empty">Select a fleet worker to show metadata and log output.</div>
+    </div>
+  </section>
+</main>
+<script type="application/json" id="fleet-initial-state">{{FLEET_INITIAL_STATE}}</script>
+<script src="/static/fleet.js"></script>
+
+</body>
+</html>
